Change type min_buy / max_buy from `int` to `float`

+++ backend/database/migrations/1617820650412_tiers_schema.js
@@ -13,7 +13,7 @@ class TiersSchema extends Schema {
       table.string('start_time').nullable();
       table.string('end_time').nullable();
       table.string('currency').nullable();
-      table.integer('max_buy').nullable().defaultTo(0);
+      table.float('max_buy').nullable().defaultTo(0);
       table.integer('campaign_id').unsigned().nullable();
 
       table.timestamps()
